refactoring 3DS file names and update features
