+++ Pagador/Transaction/Api/Auth3Ds20/Token/ResponseInterface.php
@@ -7,12 +7,11 @@
  * @link        http://www.webjump.com.br
  *
  */
-namespace Webjump\Braspag\Pagador\Transaction\Api\Auth\Token;
-
+namespace Webjump\Braspag\Pagador\Transaction\Api\Auth3Ds20\Token;
 
 interface ResponseInterface
 {
-    public function getAccessToken();
+    public function getToken();
 
     public function getTokenType();
